pyproject.toml
[project]
name = "vt-hacks-2025"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "huggingface-hub>=0.35.1",
    "mediapipe>=0.10.14",
    "instagrapi>=2.0.0",
    "mss>=10.1.0",
    "openai-whisper>=20250625",
    "opencv-python>=4.12.0.88",
    "pandas>=2.3.2",
    "pyaudio>=0.2.14",
    "pygetwindow>=0.0.9",
    "timm>=1.0.20",
    "pydantic-settings>=2.11.0",
    "openai>=1.109.1",
]

[project.optional-dependencies]
cpu = ["torch>=2.8.0", "torchvision>=0.23.0"]
cu129 = ["torch>=2.8.0", "torchvision>=0.23.0"]

[tool.uv]
conflicts = [[{ extra = "cpu" }, { extra = "cu129" }]]

[tool.uv.sources]
torch = [
    { index = "pytorch-cpu", extra = "cpu" },
    { index = "pytorch-cu129", extra = "cu129" },
]
torchvision = [
    { index = "pytorch-cpu", extra = "cpu" },
    { index = "pytorch-cu129", extra = "cu129" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu129"
url = "https://download.pytorch.org/whl/cu129"
explicit = true
